Safari-609.1.11.1/LayoutTests/platform/gtk/http/tests/websocket/tests/hybi/handshake-fail-by-null-char-in-status-expected.txt

Connection should fail immediately, rather than succeeding or staying in limbo until timeout, if a null byte is received in the status header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS timedOut is false
PASS connected is false
PASS wsOrigin is undefined.
PASS successfullyParsed is true

TEST COMPLETE
